+++ Part2/phonebook/src/App.js
@@ -1,12 +1,8 @@
 import React, {useState} from 'react';
+import Filter from './components/Filter'
+import PersonForm from './components/PersonForm'
+import Persons from './components/Persons'
 
-const Person = ({person, number}) => {
-  return(
-    <p>
-      {person} {number}
-    </p>
-  )
-}
 
 const App = () => {
   const [ persons, setPersons ] = useState([
@@ -19,75 +15,32 @@ const App = () => {
   const [newNumber, setNewNumber] = useState('phone-number')
   const [filterPerson, setFilterPerson] = useState('');
 
-  const checkPersonInArray = (person) => {
-      return person.name === newPerson
-  }
-
-  const addPerson = (event) => {
-    event.preventDefault()
-
-    if (persons.find(checkPersonInArray) !== undefined && persons.find(checkPersonInArray).name === newPerson){
-      alert(newPerson + " is already added to phonebook.");
-    }
-
-    else{
-      const personObject = {
-        name: newPerson,
-        number: newNumber
-      }
-      setPersons(persons.concat(personObject))
-      setNewPerson("")
-      setNewNumber("")
-    }
-  }
-
-  const filterPersonList = () => {
-    return persons.filter(person => 
-      person.name.slice(0, filterPerson.length).toLowerCase() === filterPerson.toLowerCase())
-  }
- 
-  const handlePersonChange = (event) => {
-    setNewPerson(event.target.value)
-  }
-
-  const handleNumberChange = (event) => {
-    setNewNumber(event.target.value)
-  }
-
-  const handlePersonFilter = (event) => {
-    setFilterPerson(event.target.value)
-  }
-
   return (
     <div>
       <h1>Phonebook</h1>
       
-      <div>
-        filter shown with <input valute={filterPerson} onChange={handlePersonFilter}/>
-      </div>
+      <Filter 
+        filterPerson={filterPerson} 
+        setFilterPerson={setFilterPerson}
+      />
+      
       <h2>add a new</h2>
-      <form onSubmit={addPerson}>
-        <div>
-          name: <input value={newPerson} onChange={handlePersonChange}/><br/><br />
-          number: <input value={newNumber} onChange={handleNumberChange}/><br/><br />
-        </div>
-        <div>
-          <button type="submit">add</button>
-        </div>
-      </form>
+      
+      <PersonForm 
+        persons={persons} 
+        newPerson={newPerson} 
+        newNumber={newNumber}
+        setPersons={setPersons} 
+        setNewPerson={setNewPerson} 
+        setNewNumber={setNewNumber}
+      />
       
       <h2>Numbers</h2>
 
-      <div>
-        {/* {persons.map(person => 
-          <Person key={person.name} person={person.name} number={person.number}/>
-        )} */}
-        {
-          filterPersonList().map(person =>
-            <Person key={person.name} person={person.name} number={person.number}/>
-            )
-        }
-      </div>
+      <Persons 
+        persons={persons} 
+        filterPerson={filterPerson} 
+      />
     </div>
   )
 }
